fix(bulk): use Long for categoryId to match jOOQ CATEGORY_ID field
BUDGETS/TRANSACTIONS.CATEGORY_ID is Field<Long?> in the generated jOOQ
code, so passing Int caused 'None of the following candidates is
applicable' on UpdateSetMoreStep.set. Propagate Long through the
bulk-categorize chain (repo, service, REST request) and the
TransactionFilters.categoryIds list.

Also fix operator-precedence bug in NotificationRepository.hasRecent:
'fetchOne(...) ?: 0 > 0' parses as 'fetchOne ?: (0 > 0)', producing a
type mismatch. Hoist the count into a local.

+++ centsible-api/src/main/kotlin/beer/thierry/centsible/api/model/transaction/TransactionFilters.kt
@@ -8,7 +8,7 @@ enum class TransactionSort {
 
 data class TransactionFilters(
     val search: String? = null,
-    val categoryIds: List<Int>? = null,
+    val categoryIds: List<Long>? = null,
     val from: LocalDate? = null,
     val to: LocalDate? = null,
     val sort: TransactionSort = TransactionSort.DATE_DESC,
